requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
python-multipart==0.0.12
httpx==0.27.0
openai==1.54.0
# playwright==1.40.0  # TODO: greenlet compatibility issue with Python 3.13
# beautifulsoup4==4.12.3  # Used with playwright for link preview
